+++ app/scripts/infuse/dimInfuse.controller.js
@@ -9,6 +9,12 @@
   function dimInfuseCtrl($scope, dimStoreService, dimItemService, ngDialog, dimLoadoutService, toaster, $q) {
     var vm = this;
 
+    if (ga) {
+      // Disable sending pageviews on popups for now, over concerns that we'll go over our free GA limits.
+      // Send a virtual pageview event, even though this is a popup
+      // ga('send', 'pageview', { page: '/infuse' });
+    }
+
     angular.extend(vm, {
       getAllItems: true,
       showLockedItems: false,