+++ app/scripts/services/dimBungieService.factory.js
@@ -104,7 +104,7 @@
 
     function openBungieNetTab() {
       chrome.tabs.create({
-        url: 'http://bungie.net',
+        url: 'https://bungie.net',
         active: true
       });
     }