+++ app/scripts/services/dimSettingsService.factory.js
@@ -35,6 +35,8 @@
       itemDetails: false,
       // Show item quality percentages
       itemQuality: false,
+      // Show new items with an overlay
+      showNewItems: false,
       // Show animation of new item overlay on new items
       showNewAnimation: true,
       // Show elemental damage icons