+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Next
+
+* New glimmer-based filters, is:glimmeritem, is:glimmerboost, is:glimmersupply
+* Add option for new item and its popup to be hidden
 * Add ability to exclude items from loadout builder.
 * Expand/collapse sections in DIM.
 * Double clicking an item will equip it on the current character. 2x click on equipped, dequips.